fix(byok): extract task taxonomy into zero-dep leaf so the web prod build resolves
The BYOK routing UI imports the value TASK_ROUTING_FALLBACK (and the LlmTask
type) from @libs/llm/byok-config. The isolated apps/web Docker build copies
libs à la carte and never included libs/llm, so `next build` failed with
"Module not found: Can't resolve '@libs/llm/byok-config'" — while local dev
(full repo on disk) passed, hiding it until the QA Deploy image build.

Move LLM_TASK / LlmTask / TASK_ROUTING_FALLBACK into a new zero-dependency
leaf libs/llm/llm-tasks.ts. byok-config re-exports them unchanged for the
backend importers; the web imports straight from the leaf. Dockerfile.web
copies just that one file — never the rest of libs/llm (AI SDK, providers,
kernel) — mirroring the existing libs/identity/domain/permissions carve-out.

Verified with a real `docker build -f docker/Dockerfile.web`: compiles the
BYOK pages and produces the full production image.
