+++ libs/llm/byok-config.ts
@@ -7,8 +7,15 @@
  * the apiKey kept as ENCRYPTED ciphertext so decryption happens once, downstream.
  * Provider ids are plain strings matching BYOKProvider values.
  */
-import { BYOKProvider } from '@libs/llm/model-providers';
+import type { BYOKProvider } from '@libs/llm/model-providers';
 import type { ReasoningEffort } from './providers/kernel/types';
+// The task taxonomy + inheritance map live in a zero-dependency leaf so the
+// isolated apps/web build can bundle them without pulling in the rest of
+// libs/llm. Re-exported here (unchanged public API) for the backend, which
+// imports them from `@libs/llm/byok-config`.
+import type { LlmTask } from './llm-tasks';
+export { LLM_TASK, TASK_ROUTING_FALLBACK } from './llm-tasks';
+export type { LlmTask } from './llm-tasks';
 
 // ─── Persisted shape ─────────────────────────────────────────────────────────
 
@@ -57,52 +64,6 @@ export interface BYOKModelConfig {
      *  sibling to rpm/tpm. */
     cooldownMs?: number;
 }
-
-/**
- * LLM task taxonomy for routing — the named source of truth for every
- * `resolveTaskSlot`/`resolveTaskModel` call. `as const` keeps the values plain
- * strings (so `routing.taskOverrides` keys, stored configs, and JSON still
- * match), while `LlmTask` is derived from it so the union can never drift from
- * the constant. Reference tasks as `LLM_TASK.codeReview`, not the bare literal.
- */
-export const LLM_TASK = {
-    codeReview: 'codeReview',
-    /** The Kody-Rules review agent (categoryLabel 'kody_rules') — enforcing the
-     *  org's authored rules on a PR. A distinct workload from the defect-finding
-     *  agents (bug/perf/security/generalist) that stay under `codeReview`, so it
-     *  can run on a different (often cheaper) model. Inherits `codeReview` when
-     *  unset (see TASK_ROUTING_FALLBACK). */
-    kodyRulesReview: 'kodyRulesReview',
-    /** Generating/learning Kody Rules (from PR history, feedback, initial scan).
-     *  Generative work, unlike per-PR detection. Inherits `codeReview` when unset. */
-    ruleGeneration: 'ruleGeneration',
-    /** The business-rules validation agent. Inherits `codeReview` when unset. */
-    businessValidation: 'businessValidation',
-    prSummary: 'prSummary',
-    conversation: 'conversation',
-} as const;
-
-export type LlmTask = (typeof LLM_TASK)[keyof typeof LLM_TASK];
-
-/**
- * Task → the task it inherits a routing target from when it has no explicit
- * `taskOverrides[task]`. The resolver applies it between the task's own override
- * and the org default (see StaticTaskStrategy); the routing UI mirrors it to
- * nest inheriting rows under their parent. Lives HERE (a bundle-safe leaf, like
- * LLM_TASK) so the backend resolver and the frontend grid import the SAME map
- * instead of hand-copying it — the inheritance graph has one source of truth.
- *
- * Each addition inherits the task it was carved out of, so an org that never
- * sets the new override keeps exactly today's behavior:
- *  - kodyRulesReview / ruleGeneration were part of the code-review flow → codeReview
- *  - businessValidation ran on the agent (chat) model → conversation
- * A single hop only (no chains).
- */
-export const TASK_ROUTING_FALLBACK: Partial<Record<LlmTask, LlmTask>> = {
-    kodyRulesReview: LLM_TASK.codeReview,
-    ruleGeneration: LLM_TASK.codeReview,
-    businessValidation: LLM_TASK.conversation,
-};
 
 /**
  * Routing policy. Persisted in Phase 2; EXECUTED in Phase 4 (Manual = static
